Refactor layout and disable slate theme
Adjusted the structure of `MainLayout.razor` by moving the closing `</div>` tag for the `top-row` div to include the `main` tag, ensuring proper nesting. Added a new `<BSCore/>` component within the `article` tag.

In `index.html`, added a comment to the line linking to the `slate` theme stylesheet, effectively disabling it to use the default Bootstrap stylesheet instead.